assignment_files/calendar/requirements.txt
Flask
requests
psycopg2-binary
flask_restful
